requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14
psycopg[binary]>=3.1
djangorestframework-simplejwt>=5.3
django-filter>=23.5
drf-spectacular>=0.27
django-cors-headers>=4.3
django-polymorphic>=3.1
gunicorn>=21.2
whitenoise>=6.6
